+++ tests/core/tests/test_resources/test_modelresource/test_data_handling.py
@@ -117,5 +117,6 @@ class DataHandlingTest(TestCase):
             {
                 "could not convert string to float",
                 "[<class 'decimal.ConversionSyntax'>]",
+                "Invalid literal for Decimal: 'foo'",
             },
         )